LibRaw 0.22.2

+++ PKGBUILD
@@ -1,17 +1,22 @@
-# Maintainer: Rodrigo Bezerra <rodrigobezerra21 at gmail dot com>
-# Contributor: orumin <dev at orum.in>
+# Maintainer: Ketchup901 <ketchup901@riseup.net>
 
-_basename=libraw
 pkgname=lib32-libraw
-pkgver=0.21.1
+pkgver=0.22.2
 pkgrel=1
 pkgdesc="A library for reading RAW files obtained from digital photo cameras (CRW/CR2, NEF, RAF, DNG, and others) (32-bit)"
 arch=('x86_64')
-url="https://www.libraw.org/"
-license=(CDDL LGPL)
-depends=(lib32-lcms2 lib32-jasper libraw)
+url='https://www.libraw.org/'
+license=(CDDL-1.0
+         LGPL-2.1-only)
+depends=(libraw
+         lib32-glibc
+         lib32-lcms2
+         lib32-gcc-libs
+         lib32-libjpeg-turbo
+         lib32-zlib)
+makedepends=(git)
 source=("https://www.libraw.org/data/LibRaw-$pkgver.tar.gz")
-sha256sums=('630a6bcf5e65d1b1b40cdb8608bdb922316759bfb981c65091fec8682d1543cd')
+sha256sums=('de86b035655accff8d4010f1a221fdf50d353cb7b1422ba26f14a0db92612cfa')
 
 build() {
     cd LibRaw-$pkgver
@@ -20,22 +25,22 @@ build() {
     export CXX='g++ -m32'
     export PKG_CONFIG_PATH='/usr/lib32/pkgconfig'
 
-    autoreconf -f
+    autoreconf -fiv
 
     ./configure \
-        --build=i686-pc-linux-gnu \
-        --prefix=/usr \
-        --libdir=/usr/lib32 \
-        --disable-examples
+    --build=i686-pc-linux-gnu \
+    --prefix=/usr \
+    --libdir=/usr/lib32 \
+    --disable-examples
 
     make
 }
 
 package() {
     cd LibRaw-$pkgver
-
     make DESTDIR="$pkgdir" install
 
-    rm -rf "${pkgdir}/usr/include"
-    rm -rf "${pkgdir}/usr/share"
+    rm -rf -- "${pkgdir:?}/usr/include"
+    rm -rf -- "${pkgdir:?}/usr/share"
 }
+
